Fix incorrect data loading
Each sample was reading a 785th item which was null, and assigning that to the next item in the array. I don't think this was a problem because that would've been overwritten by the first pixel in the next sample. Though, the last sample was trying to access out of bounds memory.